+++ yagi/src/main/java/org/raml/yagi/framework/grammar/BaseGrammar.java
@@ -201,7 +201,7 @@ public class BaseGrammar
      * @param range The valid range
      * @return The rule
      */
-    public IntegerTypeRule range(Range<Integer> range)
+    public IntegerTypeRule range(Range<Long> range)
     {
         return new IntegerTypeRule(range);
     }
